% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_accountls.R
\name{w3_accountls}
\alias{w3_accountls}
\title{w3_accountls}
\usage{
w3_accountls()
}
\value{
Character. List of accounts
}
\description{
List accounts this agent has been authorized to act on behalf of.
}
\examples{
w3_accountls()
}
