% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_access_claim.R
\name{w3_can_access_claim}
\alias{w3_can_access_claim}
\title{w3_can_access_claim}
\usage{
w3_can_access_claim()
}
\value{
Character. output of w3 can access claim
}
\description{
Claim delegated capabilities for the authorized account.
}
