% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_filecoin_info.R
\name{w3_can_filecoin_info}
\alias{w3_can_filecoin_info}
\title{w3_can_filecoin_info}
\usage{
w3_can_filecoin_info(piece_cid)
}
\arguments{
\item{piece_cid}{Character.}
}
\value{
Character. output of w3 can filecoin info
}
\description{
Get filecoin information for given PieceCid.
}
