% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_store_add.R
\name{w3_can_store_add}
\alias{w3_can_store_add}
\title{w3_can_store_add}
\usage{
w3_can_store_add(car_path)
}
\arguments{
\item{car_path}{Character.}
}
\value{
Character. output of w3 can store add
}
\description{
Store a CAR file with the service.
}
