% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_store_ls.R
\name{w3_can_store_ls}
\alias{w3_can_store_ls}
\title{w3_can_store_ls}
\usage{
w3_can_store_ls(json = FALSE, size = NULL, cursor = NULL, pre = FALSE)
}
\arguments{
\item{json}{Logical. Format as newline delimited JSON}

\item{size}{Integer. The desired number of results to return}

\item{cursor}{Character.}

\item{pre}{Logical. If true, return the page of results preceding the cursor}
}
\value{
Character. output of w3 can store ls
}
\description{
List CAR files in the current space.
}
