% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_store_rm.R
\name{w3_can_store_rm}
\alias{w3_can_store_rm}
\title{w3_can_store_rm}
\usage{
w3_can_store_rm(shard_cid)
}
\arguments{
\item{shard_cid}{Character.}
}
\value{
Character. output of w3 can store rm
}
\description{
Remove a CAR shard from the store.
}
