% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_upload_add.R
\name{w3_can_upload_add}
\alias{w3_can_upload_add}
\title{w3_can_upload_add}
\usage{
w3_can_upload_add(root_cid, shard_cid)
}
\arguments{
\item{root_cid}{Character.}

\item{shard_cid}{Character.}
}
\value{
Character. output of w3 can upload add
}
\description{
Register an upload - a DAG with the given root data CID that is stored in the given CAR shard(s), identified by CAR CIDs.
}
