% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_upload_ls.R
\name{w3_can_upload_ls}
\alias{w3_can_upload_ls}
\title{w3_can_upload_ls}
\usage{
w3_can_upload_ls(
  json = FALSE,
  shards = FALSE,
  size = NULL,
  cursor = NULL,
  pre = FALSE
)
}
\arguments{
\item{json}{Logical. Format as newline delimited JSON}

\item{shards}{Logical. Pretty print with shards in output}

\item{size}{Integer. The desired number of results to return}

\item{cursor}{Character.}

\item{pre}{Logical. If true, return the page of results preceding the cursor}
}
\value{
Character. output of w3 can upload ls
}
\description{
List uploads in the current space.
}
