% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_can_upload_rm.R
\name{w3_can_upload_rm}
\alias{w3_can_upload_rm}
\title{w3_can_upload_rm}
\usage{
w3_can_upload_rm(root_cid)
}
\arguments{
\item{root_cid}{Character.}
}
\value{
Character. output of w3 can upload rm
}
\description{
Remove an upload from the uploads listing.
}
