% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_coupon_create.R
\name{w3_coupon_create}
\alias{w3_coupon_create}
\title{w3_coupon_create}
\usage{
w3_coupon_create(
  did,
  password = NULL,
  can = NULL,
  expiration = NULL,
  output = NULL
)
}
\arguments{
\item{did}{Character.}

\item{password}{Character. Password for created coupon.}

\item{can}{Character. One or more abilities to delegate.}

\item{expiration}{Character. Unix timestamp when the delegation is no longer valid.}

\item{output}{Character. Path of file to write the exported delegation data to.}
}
\value{
Character. output of w3 coupon create
}
\description{
Create a coupon for a space
}
