% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_delegation_create.R
\name{w3_delegation_create}
\alias{w3_delegation_create}
\title{w3_delegation_create}
\usage{
w3_delegation_create(
  audience_did,
  can = NULL,
  name = NULL,
  type = NULL,
  expiration = NULL,
  output = NULL
)
}
\arguments{
\item{audience_did}{Character.}

\item{can}{Character. One or more abilities to delegate.}

\item{name}{Character. Human readable name for the audience receiving the delegation.}

\item{type}{Character. Type of the audience receiving the delegation, one of: device, app, service.}

\item{expiration}{Character. Unix timestamp when the delegation is no longer valid.}

\item{output}{Character. Path of file to write the exported delegation data to.}
}
\value{
Character. output of w3 delegation create
}
\description{
Create a delegation to the passed audience for the given abilities with the _current_ space as the resource.
}
