% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_delegation_ls.R
\name{w3_delegation_ls}
\alias{w3_delegation_ls}
\title{w3_delegation_ls}
\usage{
w3_delegation_ls(json = FALSE)
}
\arguments{
\item{json}{Logical. Format as newline delimited JSON}
}
\value{
Character. output of w3 delegation ls
}
\description{
List delegations created by this agent for others.
}
