% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_delegation_revoke.R
\name{w3_delegation_revoke}
\alias{w3_delegation_revoke}
\title{w3_delegation_revoke}
\usage{
w3_delegation_revoke(delegation_cid, proof = NULL)
}
\arguments{
\item{delegation_cid}{Character.}

\item{proof}{Character. Name of a file containing the delegation and any additional proofs needed to prove authority to revoke}
}
\value{
Character. output of w3 delegation revoke
}
\description{
Revoke a delegation by CID.
}
