% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_login.R
\name{w3_login}
\alias{w3_login}
\title{w3_login}
\usage{
w3_login(email)
}
\arguments{
\item{email}{Character. email address to authenticate with}
}
\description{
Authenticate this agent with your email address to gain access to all capabilities that have been delegated to it.
}
\details{
w3_login
}
\examples{
w3login('user@mail.com')
}
