% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_ls.R
\name{w3_ls}
\alias{w3_ls}
\title{w3_ls}
\usage{
w3_ls(json = FALSE, shards = FALSE)
}
\arguments{
\item{json}{Logical. Format as newline delimited JSON}

\item{shards}{Logical. Include shard information}
}
\value{
Character. output of w3 ls
}
\description{
List uploads in the current space
}
