% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_plan_get.R
\name{w3_plan_get}
\alias{w3_plan_get}
\title{w3_plan_get}
\usage{
w3_plan_get(email)
}
\arguments{
\item{email}{Character. email address to authenticate with}
}
\description{
:Displays plan given account is on
}
\examples{
w3_plan_get('user@mail.com')
}
