% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_proof_add.R
\name{w3_proof_add}
\alias{w3_proof_add}
\title{w3_proof_add}
\usage{
w3_proof_add(proof, json = FALSE, dry_run = FALSE)
}
\arguments{
\item{proof}{Character.}

\item{json}{Logical. Format as newline delimited JSON}

\item{dry_run}{Logical. Decode and view the proof but do not add it}
}
\value{
Character. output of w3 proof add
}
\description{
Add a proof delegated to this agent.
}
