% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_proof_ls.R
\name{w3_proof_ls}
\alias{w3_proof_ls}
\title{w3_proof_ls}
\usage{
w3_proof_ls(json = FALSE)
}
\arguments{
\item{json}{Logical. Format as newline delimited JSON}
}
\value{
Character. output of w3 proof ls
}
\description{
List proofs of capabilities delegated to this agent.
}
