% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_rm.R
\name{w3_rm}
\alias{w3_rm}
\title{w3_rm}
\usage{
w3_rm(cid, shards = FALSE)
}
\arguments{
\item{cid}{Character. CID to remove}

\item{shards}{Character. Remove all shards referenced by the upload from the store.
Use with caution and ensure other uploads do not reference the same shards.}
}
\value{
Character. output of w3 rm
}
\description{
Remove an upload from the uploads listing.
}
