% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_space_add.R
\name{w3_space_add}
\alias{w3_space_add}
\title{space add
 @description Add a space to the agent.
 @param proof Character. The proof is a CAR encoded delegation to _this_ agent.
 @return Character. output of w3 space add
 @export}
\usage{
w3_space_add(proof)
}
\description{
space add
 @description Add a space to the agent.
 @param proof Character. The proof is a CAR encoded delegation to _this_ agent.
 @return Character. output of w3 space add
 @export
}
