% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_space_create.R
\name{w3_space_create}
\alias{w3_space_create}
\title{w3_space_create}
\usage{
w3_space_create(
  name,
  no_recovery = FALSE,
  no_caution = FALSE,
  no_customer = FALSE,
  customer = NULL,
  no_account = FALSE,
  account = NULL
)
}
\arguments{
\item{name}{Character. Name of space}

\item{no_recovery}{Logical. Skips recovery key setup}

\item{no_caution}{Logical. Prints out recovery key without confirmation}

\item{no_customer}{Logical. Skip billing setup}

\item{customer}{Character. Billing account email}

\item{no_account}{Logical. Skip account setup}

\item{account}{Character. Managing account email}
}
\value{
Character. output of w3 space create
}
\description{
Create a new w3 space.
}
