% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_space_info.R
\name{w3_space_info}
\alias{w3_space_info}
\title{w3_space_info}
\usage{
w3_space_info(space = NULL, json = FALSE)
}
\arguments{
\item{space}{Character. The space to print information about.}

\item{json}{Logical. Format as newline delimited JSON}
}
\value{
Character. output of w3 space info
}
\description{
Show information about a space.
}
