% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_space_provision.R
\name{w3_space_provision}
\alias{w3_space_provision}
\title{w3_space_provision}
\usage{
w3_space_provision(
  name,
  customer = NULL,
  coupon = NULL,
  password = NULL,
  provider = NULL
)
}
\arguments{
\item{name}{Character. Name of space}

\item{customer}{Character. The email address of the billing account}

\item{coupon}{Character. Coupon URL to provision space with}

\item{password}{Character. Coupon password}

\item{provider}{Character. The storage provider to associate with this space.}
}
\value{
Character. output of w3 space provision
}
\description{
Associate space with a billing account
}
