% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_space_use.R
\name{w3_space_use}
\alias{w3_space_use}
\title{w3_space_use}
\usage{
w3_space_use(did)
}
\arguments{
\item{did}{Character.}
}
\value{
Character. output of w3 space use
}
\description{
Set the current space in use by the agent
}
