% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3_up.R
\name{w3_up}
\alias{w3_up}
\title{w3_up}
\usage{
w3_up(
  file,
  no_wrap = TRUE,
  hidden = NULL,
  car = NULL,
  json = NULL,
  verbose = NULL,
  shard_size = NULL,
  concurrent_requests = NULL
)
}
\arguments{
\item{file}{Character. File to upload}

\item{no_wrap}{Logical. Don't wrap input files with a directory.  (default false)}

\item{hidden}{Logical. Include paths that start with ".".}

\item{car}{Logical. File is a CAR file.  (default false)}

\item{json}{Logical. Format as newline delimited JSON}

\item{verbose}{Logical. Output more details.}

\item{shard_size}{Numeric. Shard uploads into CAR files of approximately this size in bytes.}

\item{concurrent_requests}{Numeric. Send up to this many CAR shards concurrently.}
}
\value{
Character. output of w3 up
}
\description{
Store a file(s) to the service and register an upload.
}
\examples{
w3_up('file.txt')
}
